Update README.md
This project is a clone of the Printify homepage, built using ReactJS and Tailwind CSS. The purpose of the project is to recreate the design and layout of the Printify homepage, incorporating key components such as a Profit Calculator and a Blog Section. This project demonstrates a responsive design that adapts well to various screen sizes, ensuring a seamless user experience across devices.